remove context
I hadn't thought (or tested) that through.  If the context expired the
io.Copy would error out and the goroutine would return.  That would
leave segCh with no readers and the loop would hang trying to write to
it.  What's even worse is that retry would cancel the context on its
return so the above scenario was guaranted to happen always.